requirements.txt
Django==2.2
django-angular==2.2
django-crispy-forms==1.7.2
django-tastypie==0.14.2
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
